Double-precision support for ResampleFilter and its dependencies.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@18363 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Pipeline/Data/avtAverageValueRF.h
@@ -84,7 +84,7 @@ class PIPELINE_API avtAverageValueRF : public avtRayFunction
     virtual            ~avtAverageValueRF();
 
     virtual void        GetRayValue(const avtRay *,
-                                    unsigned char rgb[3], float);
+                                    unsigned char rgb[3], double);
 
   protected:
     double                noSampleValue;
